#254 Fixed year cond

+++ web/controllers/UploadController.php
@@ -148,7 +148,7 @@ class UploadController extends \web\ext\Controller
             $teamName = $tr->find('.st_team', 0)->plaintext;
             $team = Team::model()->findByAttributes(array(
                 'name' => new \MongoRegex('/^' . preg_quote($teamName) . '$/i'),
-                'year' => date('Y'),
+                'year' => (int)date('Y'),
             ));
 
             // Parse tasks tries and time
